[feat] command 유효성 검사 class 화
